Move types and constants definition from centrifuge-commons lib to runtime-common
