--- libs/commons/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "centrifuge-commons"
-description = 'Primitives used by the Centrifuge chain'
-authors = ['The Centrifugians Tribe']
-edition = '2018'
-homepage = 'https://centrifuge.io'
-license = "LGPL-3.0"
-repository = 'https://github.com/centrifuge/centrifuge-chain'
-version = '2.0.0'
-
-[dependencies]
-# Substrate basic dependencies
-codec = { package = 'parity-scale-codec', version = '2.0.0', default-features = false, features = ['derive'] }
-
-# Substrate primitives
-sp-core = { git = "https://github.com/centrifuge/substrate.git", default-features = false, branch = "master" }
-sp-std = { git = "https://github.com/centrifuge/substrate", default-features = false, branch = "master" }
-
-# Centrifuge chain dependencies
-proofs = { package = "proofs", path = "../proofs", default-features = false }
-
-[features]
-default = ["std"]
-std = [
-    "codec/std",
-    "proofs/std",
-]
