+++ runtime/common/src/impls.rs
@@ -1,6 +1,7 @@
 //! Some configurable implementations as associated type for the substrate runtime.
 
 use super::*;
+use codec::{Decode, Encode};
 use core::marker::PhantomData;
 use frame_support::traits::{Currency, OnUnbalanced};
 use frame_support::weights::{
@@ -64,4 +65,35 @@ impl WeightToFeePolynomial for WeightToFee {
 			degree: 1,
 		})
 	}
+}
+
+/// A global identifier for an nft/asset on-chain. Composed of a registry and token id.
+#[derive(Encode, Decode, Clone, PartialEq, Eq, Default, Debug)]
+pub struct AssetId(pub RegistryId, pub TokenId);
+
+/// Holds references to its component parts.
+pub struct AssetIdRef<'a>(pub &'a RegistryId, pub &'a TokenId);
+
+impl AssetId {
+	pub fn destruct(self) -> (RegistryId, TokenId) {
+		(self.0, self.1)
+	}
+}
+
+impl<'a> From<&'a AssetId> for AssetIdRef<'a> {
+	fn from(id: &'a AssetId) -> Self {
+		AssetIdRef(&id.0, &id.1)
+	}
+}
+
+impl<'a> AssetIdRef<'a> {
+	pub fn destruct(self) -> (&'a RegistryId, &'a TokenId) {
+		(self.0, self.1)
+	}
+}
+
+/// All data for an instance of an NFT.
+#[derive(Encode, Decode, Clone, PartialEq, Eq, Default, Debug)]
+pub struct AssetInfo {
+	pub metadata: Bytes,
 }
